Simplify link and clean printing
There is no need to print every single object on a new line because the
linker is not linking incrementally.  Similarly for the artifacts, the
entire folder is deleted, so why bother printing the entire tree (which
can become very large).  The focus here is on speed and this is
unnecessary.  The compiler prints will stay on a per file basis because
the compilation is incremental, so it's useful for detected which file
the compiler issues appear at.